+++ common_pyutil/decorators.py
@@ -48,7 +48,7 @@ class Tag:
         return self._members
 
     def add(self, f: Callable):
-        """Add a function `f` to members
+        """Add a function :code:`f` to members
 
         Args:
             f: Function to add
@@ -59,6 +59,21 @@ class Tag:
                 self._members[f.fget.__name__] = f
             else:
                 self._members[f.__name__] = f
+
+    def remove(self, fname: str):
+        """Remove a function named :code:`fname` to members
+
+        Args:
+            f: Function to add
+
+        """
+        if fname in self._members:
+            self._members.pop(fname)
+        else:
+            raise AttributeError(f"No such member {fname}")
+
+    def __iter__(self):
+        return self._members.__iter__()
 
     def __call__(self, f: Callable) -> Callable:
         """Add a function `f` to members self being called as a decorator.